+++ packages/param/src/worklet.ts
@@ -19,19 +19,15 @@ export class ParamProcessor extends AudioWorkletProcessor {
     };
   }
 
-  process(
-    inputs: Float32Array[][],
-    outputs: Float32Array[][],
-    parameters: any
-  ) {
-    if (this.s !== parameters.scale[0]) {
-      this.s = parameters.scale[0];
+  process(inputs: Float32Array[][], outputs: Float32Array[][], params: any) {
+    if (this.s !== params.scale[0]) {
+      this.s = params.scale[0];
       this.c = getConverter(this.s);
     }
 
-    const input = parameters.input[0] + parameters.mod[0];
-    const scaled = this.c(input, parameters.min[0], parameters.max[0]);
-    const out = scaled * parameters.gain[0] + parameters.offset[0];
+    const input = params.input[0] + params.mod[0];
+    const scaled = this.c(input, params.min[0], params.max[0]);
+    const out = scaled * params.gain[0] + params.offset[0];
     outputs[0][0].fill(out);
 
     return this.r;