+++ packages/synthlet/CHANGELOG.md
@@ -1,11 +1,21 @@
 # synthlet
 
+## 0.7.0
+
+### Minor Changes
+
+- - Add `detune` parameter to Polyblep oscillator
+
+### Patch Changes
+
+- Updated dependencies
+  - @synthlet/polyblep-oscillator@0.2.0
+
 ## 0.6.0
 
 ### Minor Changes
 
-- d9ae8bf: State Variable Filter improvements:
-
+- State Variable Filter improvements:
   - It uses a better algorithm (by Andrew Simper and Freq Anton Corvest)
   - `frequency` is now a a-rate parameter suitable for modulation
   - Renamed `resonance` to `Q` to match Web Audio API standard